requirements.txt
streamlit
google-generativeai
huggingface-hub
python-dotenv
Pillow